+++ src/Ray.BiliBiliTool.Agent/BiliBiliAgent/Interfaces/ILiveApi.cs
@@ -78,7 +78,7 @@ namespace Ray.BiliBiliTool.Agent.BiliBiliAgent.Interfaces
         [Header("Referer", "https://live.bilibili.com/")]
         [Header("Origin", "https://live.bilibili.com")]
         [HttpGet("/xlive/web-interface/v1/second/getList?platform=web&parent_area_id={parentAreaId}&area_id={areaId}&sort_type={sortType}&page={page}")]
-        Task<BiliApiResponse<GetListResponse>> GetList(int parentAreaId, int page, int areaId = 0, string sortType = "");
+        Task<BiliApiResponse<GetListResponse>> GetList(long parentAreaId, int page, int areaId = 0, string sortType = "");
         //todo:Cookie比nav接口多了两项：Hm_lvt_8a6e55dbd2870f0f5bc9194cddf32a02、Hm_lvt_9e2a88dc69e0e55c353597501d2a4bbc
 
         /// <summary>
